// src/ui/templates/dashboard/shortcutsMenu/ShortcutsMenu.types.ts
export enum ShortcutsItem {
  PrintBrochure = 'PrintBrochure',
  DailyNvmMutations = 'DailyNvmMutations',
  GenerateContract = 'GenerateContract',
  DashboardAllocate = 'DashboardAllocate',
  WwftCheck = 'WwftCheck',
  Allocate = 'Allocate',
  Match = 'Match',
  DashboardKpi = 'DashboardKpi',
  SendNewsletter = 'SendNewsletter',
}